refactor: cardRegisterInputProps를 components/fields/shared.types로 이동 및 import 경로 반영
