+++ src/feature/CardRegister/components/fields/CvcField/CvcField.stories.tsx
@@ -4,7 +4,7 @@ import type {Meta, StoryObj} from '@storybook/react-vite';
 import {fn} from 'storybook/test';
 
 import CvcField from './CvcField';
-import type {CardRegisterInputProps} from '@/feature/CardRegister/types/cardRegisterInputProps';
+import type {CardRegisterInputProps} from '../shared.types';
 
 const createInputProps = (value: string): CardRegisterInputProps => ({
   value,
