+++ src/feature/CardRegister/components/fields/NumberField/NumberField.tsx
@@ -1,8 +1,8 @@
 import styled from 'styled-components';
 
 import Input from '@/common/components/Input/Input';
-import type {CardRegisterInputProps} from '@/feature/CardRegister/types/cardRegisterInputProps';
-import {useInputFocusMove} from '@/feature/CardRegister/hooks/useInputFocusMove';
+import type {CardRegisterInputProps} from '../shared.types';
+import {useInputFocusMove} from '@/feature/CardRegister/hooks/ui/useInputFocusMove';
 import InputContainer from '../InputContainer/InputContainer';
 import {FieldErrorMessage, FieldLayout} from '../../styles/inputFieldStyles';
